# Conflicts:
#	resto-shop-api/src/main/java/com/resto/shop/web/model/Order.java
#	resto-shop-api/src/main/java/com/resto/shop/web/model/OrderItem.java
#	resto-shop-api/src/main/java/com/resto/shop/web/service/ArticleService.java
#	resto-shop-api/src/main/java/com/resto/shop/web/service/OrderItemService.java
#	resto-shop-api/src/main/java/com/resto/shop/web/service/UnitService.java
#	resto-shop-server/src/main/java/com/resto/shop/web/aspect/OrderAspect.java
#	resto-shop-server/src/main/java/com/resto/shop/web/dao/ArticleMapper.java
#	resto-shop-server/src/main/java/com/resto/shop/web/dao/ArticleMapper.xml
#	resto-shop-server/src/main/java/com/resto/shop/web/dao/OrderItemMapper.java
#	resto-shop-server/src/main/java/com/resto/shop/web/dao/OrderItemMapper.xml
#	resto-shop-server/src/main/java/com/resto/shop/web/producer/MQMessageProducer.java
#	resto-shop-server/src/main/java/com/resto/shop/web/service/impl/ArticleServiceImpl.java
#	resto-shop-server/src/main/java/com/resto/shop/web/service/impl/OrderItemServiceImpl.java
#	resto-shop-server/src/main/java/com/resto/shop/web/service/impl/OrderServiceImpl.java
#	resto-shop-web/src/main/java/com/resto/shop/web/controller/business/ShopDetailManageController.java

+++ resto-shop-web/src/main/java/com/resto/shop/web/controller/business/ArticleController.java
@@ -11,10 +11,7 @@ import com.resto.brand.web.service.PlatformService;
 import com.resto.brand.web.service.ShopDetailService;
 import com.resto.shop.web.constant.ArticleType;
 import com.resto.shop.web.controller.GenericController;
-import com.resto.shop.web.model.Article;
-import com.resto.shop.web.model.ArticlePrice;
-import com.resto.shop.web.model.ArticleRecommendPrice;
-import com.resto.shop.web.model.MealItem;
+import com.resto.shop.web.model.*;
 import com.resto.shop.web.service.*;
 import com.resto.shop.web.util.LogTemplateUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -67,6 +64,12 @@ public class ArticleController extends GenericController {
 
     @Autowired
     private PlatformService platformService;
+
+    @Resource
+    RecommendCategoryArticleService recommendCategoryArticleService;
+
+    @Resource
+    RecommendCategoryService recommendCategoryService;
 
     @RequestMapping("/list")
     public void list() {
@@ -183,6 +186,15 @@ public class ArticleController extends GenericController {
                 result.setStatusCode(100);
                 return result;
             }
+        }
+        RecommendCategoryArticle recommendCategoryArticle=recommendCategoryArticleService.selectByArticleId(id);
+        if(recommendCategoryArticle!=null){
+            RecommendCategory recommendCategory=recommendCategoryService.selectById(recommendCategoryArticle.getRecommendCategoryId());
+            Result result = new Result();
+            result.setSuccess(false);
+            result.setMessage("删除失败，该菜品在推荐类别“" +recommendCategory.getName()+ "”中存在，请在推荐类别中删除后重试");
+            result.setStatusCode(100);
+            return result;
         }
         articleService.delete(id);
         //联动删除在推荐餐品包中的id
